@@ -8,22 +8,30 @@
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
 
-    <name>spring-boot-elasticsearch-over-http</name>
-    <description>Demo project for Spring Boot</description>
+    <name>testcontainers-junit5</name>
+    <description>Simple demo demonstrates using Spring Boot, Kotlin, Elasticsearch, JUnit5 and testcontainers</description>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.0.M2</version>
+        <version>2.1.3.RELEASE</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-        <java.version>1.8</java.version>
-        <kotlin.version>1.2.61</kotlin.version>
-        <elasticsearch-rest-high-level-client.version>6.3.2</elasticsearch-rest-high-level-client.version>
+
+        <maven.min.version>3.2.1</maven.min.version>
+        <java.version>11</java.version>
+        <kotlin.version>1.3.21</kotlin.version>
+
+        <elasticsearch.version>6.6.1</elasticsearch.version>
+
+        <!-- testing -->
+        <junit-jupiter.version>5.4.0</junit-jupiter.version>
+        <mockito-kotlin.version>2.1.0</mockito-kotlin.version>
+        <testcontainers.version>1.10.6</testcontainers.version>
     </properties>
 
     <dependencies>
