@@ -1,10 +1,10 @@
-package de.larmic.springbootelasticsearchoverhttp
+package de.larmic.testcontainers
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
 @SpringBootApplication
-class SpringBootElasticsearchOverHttpApplication
+open class SpringBootElasticsearchOverHttpApplication
 
 fun main(args: Array<String>) {
     runApplication<SpringBootElasticsearchOverHttpApplication>(*args)
