@@ -1,15 +1,17 @@
-package de.larmic.springbootelasticsearchoverhttp
+package de.larmic.testcontainers.elasticsearch
 
 import com.fasterxml.jackson.databind.DeserializationFeature
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
-import de.larmic.springbootelasticsearchoverhttp.TweetDocument.Companion.documentIndex
-import de.larmic.springbootelasticsearchoverhttp.TweetDocument.Companion.documentType
+import de.larmic.testcontainers.elasticsearch.TweetDocument.Companion.documentIndex
+import de.larmic.testcontainers.elasticsearch.TweetDocument.Companion.documentType
+import de.larmic.testcontainers.rest.Tweet
 import org.elasticsearch.action.delete.DeleteRequest
 import org.elasticsearch.action.get.GetRequest
 import org.elasticsearch.action.index.IndexRequest
 import org.elasticsearch.action.search.SearchRequest
 import org.elasticsearch.action.update.UpdateRequest
+import org.elasticsearch.client.RequestOptions
 import org.elasticsearch.client.RestHighLevelClient
 import org.elasticsearch.common.xcontent.XContentType
 import org.elasticsearch.rest.RestStatus
@@ -27,12 +29,12 @@ class TweetRepository(private val restHighLevelClient: RestHighLevelClient) {
         val writeValueAsString = mapper.writeValueAsString(document)
         val request = IndexRequest(TweetDocument.documentIndex, TweetDocument.documentType)
         request.source(writeValueAsString, XContentType.JSON)
-        return restHighLevelClient.index(request).id
+        return restHighLevelClient.index(request, RequestOptions.DEFAULT).id
     }
 
     fun getTweet(id: String): Tweet? {
         val getRequest = GetRequest(TweetDocument.documentIndex, TweetDocument.documentType, id)
-        val getResponse = restHighLevelClient.get(getRequest)
+        val getResponse = restHighLevelClient.get(getRequest, RequestOptions.DEFAULT)
         return when (getResponse.isExists) {
             true -> {
                 val sourceAsString = getResponse.sourceAsString
@@ -47,7 +49,7 @@ class TweetRepository(private val restHighLevelClient: RestHighLevelClient) {
         // see https://www.elastic.co/guide/en/elasticsearch/client/java-rest/6.x/java-rest-high-search.html
         val searchRequest = SearchRequest(TweetDocument.documentIndex)
         searchRequest.types(TweetDocument.documentType)
-        return restHighLevelClient.search(searchRequest).hits
+        return restHighLevelClient.search(searchRequest, RequestOptions.DEFAULT).hits
                 .map { Tweet(it.id, mapper.readValue(it.sourceAsString, TweetDocument::class.java).message) }
     }
 
@@ -59,13 +61,13 @@ class TweetRepository(private val restHighLevelClient: RestHighLevelClient) {
         val jsonMap = HashMap<String, Any>()
         jsonMap["message"] = tweet
         val request = UpdateRequest(TweetDocument.documentIndex, TweetDocument.documentType, id).doc(jsonMap, XContentType.JSON)
-        val updateResponse = restHighLevelClient.update(request)
+        val updateResponse = restHighLevelClient.update(request, RequestOptions.DEFAULT)
         return updateResponse.status()
     }
 
     fun deleteTweet(id: String): RestStatus {
         val request = DeleteRequest(TweetDocument.documentIndex, TweetDocument.documentType, id)
-        val deleteResponse = restHighLevelClient.delete(request)
+        val deleteResponse = restHighLevelClient.delete(request, RequestOptions.DEFAULT)
         return deleteResponse.status()
     }
 
@@ -73,6 +75,6 @@ class TweetRepository(private val restHighLevelClient: RestHighLevelClient) {
         val getRequest = GetRequest(documentIndex, documentType, id)
         getRequest.fetchSourceContext(FetchSourceContext(false))
         getRequest.storedFields("_none_")
-        return restHighLevelClient.exists(getRequest)
+        return restHighLevelClient.exists(getRequest, RequestOptions.DEFAULT)
     }
 }
